Remove rendundant and unused code in docs building/publishing (#36346)
Currently docs building happens insid of the container image and code
doing that sits in `docs` folder, while publishing has already been
moved to `breeze` code (and is executed in the Breeze venv, not in the
container). Both building and publishing code were present in both
(copy&pasted) and the parts of it not relevant to the `other` function
has not been used.

While eventually we will move docs building also to `breeze` the first
step of that is to remove the redundancy and clean-up unused code, so
that we can make the transition cleaner.
