+++ docs/exts/docs_build/docs_builder.py
@@ -26,13 +26,11 @@ from subprocess import run
 from rich.console import Console
 
 from .code_utils import (
-    AIRFLOW_SITE_DIR,
     ALL_PROVIDER_YAMLS,
     ALL_PROVIDER_YAMLS_WITH_SUSPENDED,
     CONSOLE_WIDTH,
     DOCS_DIR,
     PROCESS_TIMEOUT,
-    pretty_format_path,
 )
 from .errors import DocBuildError, parse_sphinx_warnings
 from .helm_chart_utils import chart_version
@@ -104,13 +102,6 @@ class AirflowDocsBuilder:
         if self.package_name == "helm-chart":
             return chart_version()
         return Exception(f"Unsupported package: {self.package_name}")
-
-    @property
-    def _publish_dir(self) -> str:
-        if self.is_versioned:
-            return f"docs-archive/{self.package_name}/{self._current_version}"
-        else:
-            return f"docs-archive/{self.package_name}"
 
     @property
     def _src_dir(self) -> str:
@@ -271,31 +262,6 @@ class AirflowDocsBuilder:
         else:
             console.print(f"[info]{self.package_name:60}:[/] [green]Finished docs building successfully[/]")
         return build_errors
-
-    def publish(self, override_versioned: bool):
-        """Copy documentation packages files to airflow-site repository."""
-        console.print(f"Publishing docs for {self.package_name}")
-        output_dir = os.path.join(AIRFLOW_SITE_DIR, self._publish_dir)
-        pretty_source = pretty_format_path(self._build_dir, os.getcwd())
-        pretty_target = pretty_format_path(output_dir, AIRFLOW_SITE_DIR)
-        console.print(f"Copy directory: {pretty_source} => {pretty_target}")
-        if os.path.exists(output_dir):
-            if self.is_versioned:
-                if override_versioned:
-                    console.print(f"Overriding previously existing {output_dir}! ")
-                else:
-                    console.print(
-                        f"Skipping previously existing {output_dir}! "
-                        f"Delete it manually if you want to regenerate it!"
-                    )
-                    console.print()
-                    return
-            shutil.rmtree(output_dir)
-        shutil.copytree(self._build_dir, output_dir)
-        if self.is_versioned:
-            with open(os.path.join(output_dir, "..", "stable.txt"), "w") as stable_file:
-                stable_file.write(self._current_version)
-        console.print()
 
 
 def get_available_providers_packages(include_suspended: bool = False):
